Pausing Rework
- Implemented a massive change to how pausing work, removing displaying surfaces and instead taking the application_surface and turning it into a new background sprite to display.

This change has allowed for combat pause cameras to work as intended and for the implementation of the main menu.

This solution still needs work, as only the application surface is displayed when paused, meaning any camera movement will show black spaces.

- Implemented basic ESC menu allowing for resuming, quitting game, and returning to main menu.

Returning to main menu does not work as intended and has some strange interactions.